pyproject.toml
[tool.poetry]
name = "myrmidon"
version = "0.1.0"
description = "General purpose task runner bot for the CONsortium Discord server."
authors = ["Benjamin Herman <benjamin@metanomial.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
hikari = "^2.0.0.dev115"
python = "~3.11"
python-dotenv = "^0.21.0"
schedule = "^1.1.0"
requests = "^2.28.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
pre-commit = "^2.21.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
